General corrections of warnings (-90%) & files format

+++ Camera/CameraPike/camera_pike.cc
@@ -113,7 +113,7 @@ namespace camera
 
     desc_.width = pt.get<unsigned short>("pike.sensor_width", desc_.width);
     desc_.height = pt.get<unsigned short>("pike.sensor_height", desc_.height);
-    desc_.depth = (pt.get<unsigned short>("pike.bit_depth", 8) + 7) / 8;
+    desc_.depth = static_cast<float>(pt.get<unsigned short>("pike.bit_depth", 8) + 7) / 8.f;
 
     subsampling_ = pt.get<int>("pike.subsampling", subsampling_);
     gain_ = pt.get<unsigned long>("pike.gain", gain_);
